Updated the update.sh and include versioncheck.py

+++ update.sh
@@ -5,16 +5,21 @@ read gitDownload
 echo
 if [ "$gitDownload" == "y" ];
 then
-
-    read -p "USERNAME: " USERNAME
-    read -sp "PASSWORD: " PASSWORD
-    cd ..
-    
-    rm -rf CommandConsole
-    git clone https://github.com/Shreejan-35/CommandConsole.git
-    cd CommandConsole
-	echo $USERNAME > etc/userdetails/user.txt
-	echo $PASSWORD > etc/userdetails/passwd.txt
+    python etc/version/versioncheck.py
+    if [ -e etc/version/newVersion.txt ]
+    then
+        read -p "USERNAME: " USERNAME
+        read -sp "PASSWORD: " PASSWORD
+        cd ..
+        
+        rm -rf CommandConsole
+        git clone https://github.com/Shreejan-35/CommandConsole.git
+        cd CommandConsole
+        echo $USERNAME > etc/userdetails/user.txt
+        echo $PASSWORD > etc/userdetails/passwd.txt
+    else
+        echo "bye"
+    fi
 else    
     echo "Open any browser and install git, you can directly head over to https://git-scm.com/downloads"
     echo "After downloading and initializing if you want, you can again start this script"
